Acordion 1 moved to event page
